fix: AccountData を types.ts に移動して循環依存を解消する
types.ts が AccountSpreadsheet から AccountData を import し、
AccountSpreadsheet が types.ts から RoleLabel を import する
lib → UI → lib の循環依存を解消する。
AccountData の定義を types.ts に移動し、AccountSpreadsheet は
re-export のみ行う形に変更。依存の向きを UI → lib に統一する。
