+++ src/app/page.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { useCallback, useState } from "react";
-import type { AccountData } from "@/components/AccountSpreadsheet";
 import AccountSpreadsheet from "@/components/AccountSpreadsheet";
 import { generateMembersSql } from "@/lib/sql/generateMembersSql";
 import { generateUsersSql } from "@/lib/sql/generateUsersSql";
 import { ROLE_MAP } from "@/lib/sql/types";
+import type { AccountData } from "@/lib/sql/types";
 
 // bcryptjs is used synchronously here (existing behavior). For large workloads consider
 // moving hashing into a Web Worker or to the server.
